+++ src/Symfony/Bridge/Twig/Tests/Mime/TemplatedEmailTest.php
@@ -17,6 +17,8 @@ use Symfony\Component\Mime\Part\DataPart;
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainNormalizer;
 use Symfony\Component\Serializer\Normalizer\MimeMessageNormalizer;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Normalizer\PropertyNormalizer;
@@ -105,12 +107,23 @@ EOF;
 
         $extractor = new PhpDocExtractor();
         $propertyNormalizer = new PropertyNormalizer(null, null, $extractor);
-        $serializer = new Serializer([
+        $normalizers = [
+            new MimeMessageNormalizer($propertyNormalizer),
+            new ObjectNormalizer(null, null, null, $extractor),
+            $propertyNormalizer,
+        ];
+        $denormalizers = [
             new ArrayDenormalizer(),
             new MimeMessageNormalizer($propertyNormalizer),
             new ObjectNormalizer(null, null, null, $extractor),
             $propertyNormalizer,
-        ], [new JsonEncoder()]);
+        ];
+        $serializer = new Serializer(
+            [],
+            [new JsonEncoder()],
+            new ChainNormalizer($normalizers),
+            new ChainDenormalizer($denormalizers),
+        );
 
         $serialized = $serializer->serialize($e, 'json', [ObjectNormalizer::IGNORED_ATTRIBUTES => ['cachedBody']]);
         $this->assertStringMatchesFormat($expectedJson, json_encode(json_decode($serialized), \JSON_PRETTY_PRINT | \JSON_UNESCAPED_SLASHES));