+++ src/Symfony/Component/Serializer/Debug/TraceableEncoder.php
@@ -15,6 +15,8 @@ use Symfony\Component\Serializer\DataCollector\SerializerDataCollector;
 use Symfony\Component\Serializer\Encoder\DecoderInterface;
 use Symfony\Component\Serializer\Encoder\EncoderInterface;
 use Symfony\Component\Serializer\Encoder\NormalizationAwareInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Serializer\SerializerAwareInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 
@@ -25,7 +27,7 @@ use Symfony\Component\Serializer\SerializerInterface;
  *
  * @internal
  */
-class TraceableEncoder implements EncoderInterface, DecoderInterface, SerializerAwareInterface
+class TraceableEncoder implements EncoderInterface, DecoderInterface, SerializerAwareInterface, NormalizerAwareInterface
 {
     public function __construct(
         private EncoderInterface|DecoderInterface $encoder,
@@ -92,6 +94,15 @@ class TraceableEncoder implements EncoderInterface, DecoderInterface, Serializer
         }
 
         $this->encoder->setSerializer($serializer);
+    }
+
+    public function setNormalizer(NormalizerInterface $normalizer): void
+    {
+        if (!$this->encoder instanceof NormalizerAwareInterface) {
+            return;
+        }
+
+        $this->encoder->setNormalizer($normalizer);
     }
 
     public function needsNormalization(): bool