+++ src/Symfony/Component/Serializer/Tests/Normalizer/PropertyNormalizerTest.php
@@ -24,6 +24,8 @@ use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter
 use Symfony\Component\Serializer\NameConverter\MetadataAwareNameConverter;
 use Symfony\Component\Serializer\Normalizer\AbstractObjectNormalizer;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainNormalizer;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\PropertyNormalizer;
 use Symfony\Component\Serializer\Serializer;
@@ -251,7 +253,7 @@ class PropertyNormalizerTest extends TestCase
     protected function getNormalizerForCircularReference(array $defaultContext): PropertyNormalizer
     {
         $normalizer = new PropertyNormalizer(null, null, null, null, null, $defaultContext);
-        new Serializer([$normalizer]);
+        new Serializer([], [], $normalizer);
 
         return $normalizer;
     }
@@ -268,7 +270,7 @@ class PropertyNormalizerTest extends TestCase
 
     public function testSiblingReference()
     {
-        $serializer = new Serializer([$this->normalizer]);
+        $serializer = new Serializer([], [], $this->normalizer);
         $this->normalizer->setSerializer($serializer);
 
         $siblingHolder = new PropertySiblingHolder();
@@ -285,7 +287,7 @@ class PropertyNormalizerTest extends TestCase
     {
         $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
         $denormalizer = new PropertyNormalizer($classMetadataFactory, new MetadataAwareNameConverter($classMetadataFactory));
-        $serializer = new Serializer([$denormalizer]);
+        $serializer = new Serializer([], [], null, new ChainDenormalizer([$denormalizer]));
         $denormalizer->setSerializer($serializer);
 
         return $denormalizer;
@@ -353,7 +355,7 @@ class PropertyNormalizerTest extends TestCase
     {
         $normalizer = new PropertyNormalizer();
         // instantiate a serializer with the normalizer to handle normalizing recursive structures
-        new Serializer([$normalizer]);
+        new Serializer([], [], null, $normalizer);
 
         return $normalizer;
     }
@@ -362,7 +364,7 @@ class PropertyNormalizerTest extends TestCase
     {
         $normalizer = new PropertyNormalizer();
         // instantiate a serializer with the normalizer to handle normalizing recursive structures
-        new Serializer([$normalizer]);
+        new Serializer([], [], $normalizer);
 
         return $normalizer;
     }
@@ -376,7 +378,7 @@ class PropertyNormalizerTest extends TestCase
     {
         $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
         $normalizer = new PropertyNormalizer($classMetadataFactory);
-        $serializer = new Serializer([$normalizer]);
+        $serializer = new Serializer([], [], $normalizer);
         $normalizer->setSerializer($serializer);
 
         return $normalizer;
@@ -386,7 +388,7 @@ class PropertyNormalizerTest extends TestCase
     {
         $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
         $normalizer = new PropertyNormalizer($classMetadataFactory, null, new PhpDocExtractor());
-        new Serializer([$normalizer]);
+        new Serializer([], [], null, $normalizer);
 
         return $normalizer;
     }
@@ -395,7 +397,7 @@ class PropertyNormalizerTest extends TestCase
     {
         $extractor = new PropertyInfoExtractor([], [new PhpDocExtractor(), new ReflectionExtractor()]);
         $normalizer = new PropertyNormalizer(null, null, $extractor);
-        $serializer = new Serializer([new ArrayDenormalizer(), $normalizer]);
+        $serializer = new Serializer([], [], null, new ChainDenormalizer([new ArrayDenormalizer(), $normalizer]));
         $normalizer->setSerializer($serializer);
 
         return $normalizer;